Group routes for SpendingController as a resource

+++ app/Http/Controllers/SpendingController.php
@@ -50,6 +50,8 @@ class SpendingController extends Controller {
     }
 
     public function destroy(Spending $spending) {
+        $this->authorize('delete', $spending);
+
         $spending->delete();
 
         return redirect()->route('spendings.index');
